Develop

+++ db/migrate/20240603151938_create_posts.rb
@@ -3,7 +3,7 @@ class CreatePosts < ActiveRecord::Migration[6.1]
     create_table :posts do |t|
 
       t.timestamps
-      t.integer :user_id, null: false
+      t.bigint :user_id, null: false
       t.string :title, null: false
       t.text :body, null: false
     end
