+++ db/migrate/20240603151959_create_comments.rb
@@ -3,8 +3,8 @@ class CreateComments < ActiveRecord::Migration[6.1]
     create_table :comments do |t|
 
       t.timestamps
-      t.integer :user_id, null: false
-      t.integer :post_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :post_id, null: false
       t.text :comment, null: false
     end
   end
